agent/requirements.txt
google-adk
mcp
httpx
anyio
